+++ comments-api/src/main/java/com/example/commentsapi/mq/Receiver.java
@@ -1,16 +1,21 @@
 package com.example.commentsapi.mq;
 
+import com.example.commentsapi.service.CommentServiceImpl;
 import org.springframework.amqp.rabbit.annotation.RabbitHandler;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @Component
 @RabbitListener(queues = "queue1")
 public class Receiver {
+    @Autowired
+    private CommentServiceImpl commentService;
 
     @RabbitHandler
     public void receive(String msg) throws InterruptedException {
-        System.out.println("message received: " + msg);
+        System.out.println("postId received to delete: " + msg);
+        commentService.deleteByPostId(Integer.parseInt(msg));
     }
 
 //    @RabbitHandler
