+++ posts-api/src/main/resources/application-dev.properties
@@ -13,3 +13,9 @@ spring.datasource.username=postgres
 spring.datasource.password=password
 
 spring.jpa.properties.hibernate.temp.use_jdbc_metadata_defaults=false
+
+spring.rabbitmq.host=rabbitmq
+spring.rabbitmq.port=5672
+spring.rabbitmq.username=guest
+spring.rabbitmq.password=guest
+spring.rabbitmq.dynamic=true
